move files around so that the tests live near the logic they're testing
