+++ src/App.test.tsx
@@ -1,4 +1,4 @@
-import App from "../App";
+import App from "./App";
 import { render, screen } from "@testing-library/react";
 import { describe, it, expect } from "vitest";
